telegram_bot_project/pyproject.toml
[tool.poetry]
name = "telegram-groups-bot"
version = "1.0.0"
description = "بوت تيليجرام لإنشاء المجموعات تلقائياً"
author = "Manus AI"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "20.7"
telethon = "1.34.0"
asyncio-mqtt = "0.16.1"
aiofiles = "23.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

